Do nothing if there is no matching record in Reaction.destroy_account

+++ app/models/reaction.rb
@@ -43,6 +43,8 @@ class Reaction < ApplicationRecord
       ApplicationRecord.transaction isolation: :repeatable_read do
         reaction = find_by!(attributes)
 
+        next unless reaction.accounts.exists? account.id
+
         if reaction.accounts_count <= 1
           reaction.destroy!
         else
